feat: refactor to keystore v4
