+++ projects/wallet-template/src/background/keystore/index.ts
@@ -1,0 +1,1 @@
+export * as keystoreV4 from "./keystoreV4"
